+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v24.0.0...master)
+[Full changelog](https://github.com/mozilla/glean/compare/v24.1.0...master)
+
+# v24.1.0 (2020-01-16)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v24.0.0...v24.1.0)
+
+* Android:
+  * A new metric `glean.error.preinit_tasks_overflow` was added to report when
+    the preinit task queue overruns, leading to data loss. See [bug
+    1609482](https://bugzilla.mozilla.org/show_bug.cgi?id=1609482)
 
 # v24.0.0 (2020-01-14)